pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-invoice-agent"
version = "0.1.0"
description = "Frontend plus local worker for an AI invoice agent."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "playwright>=1.52.0",
  "pypdf>=5.0.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
